https/archive/https/requirements.txt
appdirs==1.4.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
idna==3.4
lxml==4.9.2
Naked==0.1.32
packaging==23.0
py==1.11.0
pycparser==2.21
pycryptodome==3.17
pyshark==0.5.3
PyYAML==6.0
requests==2.28.2
shellescape==3.8.1
urllib3==1.26.14
